LeSeminaire/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "imageio>=2.37.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "yt-dlp>=2025.3.31",
    "werkzeug>=3.1.3",
    "slack-sdk>=3.35.0",
    "psutil>=7.0.0",
]
